+++ source/index.ts
@@ -1,49 +1,84 @@
 /* eslint-disable no-use-before-define */
 
 // external
-import fetch from 'node-fetch'
-import Errlop from 'errlop'
 import withinRange from 'version-range'
 import versionCompare from 'version-compare'
-import { getDateWithYearOffset, getESVersion } from 'es-versions'
 import { last } from '@bevry/list'
-export { last } from '@bevry/list'
 
-/** A version number in string or number format  */
-export type Input = string | number
+// related
+import {
+	NodeScheduleIdentifier,
+	preloadNodeSchedule,
+	getNodeScheduleIdentifiers,
+	getNodeScheduleInformation,
+	NodeScheduleInput,
+	NodeScheduleInformation,
+} from '@bevry/nodejs-schedule'
+import {
+	NodeReleaseInput,
+	preloadNodeReleases,
+	getNodeReleaseInformation,
+	NodeReleaseIdentifier,
+	getNodeReleaseIdentifiers,
+} from '@bevry/nodejs-releases'
 
-/** A version number in desired string format */
-export type Key = string
+// ====================================
+// API
 
-/** The metadata for a version */
-export interface Meta {
-	version: string
-	start: Date
-	end: Date
-	lts?: Date
-	maintenance?: Date
-	codename?: string
+/** Either a significant or absolute Node.js version number input */
+export type NodeVersionInput = NodeScheduleInput | NodeReleaseInput
+
+/** Either a significant or absolute Node.js version number */
+export type NodeVersionIdentifier =
+	| NodeScheduleIdentifier
+	| NodeReleaseIdentifier
+
+/** Helper method to detect if a version number is an absolute version number or not. */
+export function isAbsoluteVersion(version: NodeVersionInput): boolean {
+	return String(version).split('.').length === 3
 }
 
-/** The response format from https://github.com/nodejs/Release */
-interface Response {
-	[version: string]: {
-		start: string
-		end: string
-		lts?: string
-		maintenance?: string
-		codename?: string
+/** Helper to return `false` if the Node.js version does not exist in the schedule (prior to 0.8, or say 0.9, or 0.11) */
+export function getNodeScheduleInformationSafe(
+	version: NodeScheduleInput
+): NodeScheduleInformation | false {
+	if (
+		versionCompare(version, '0.8') === -1 ||
+		versionCompare(version, '0.9') === 0 ||
+		versionCompare(version, '0.11') === 0
+	) {
+		return false
 	}
+	return getNodeScheduleInformation(version)
 }
 
 // prepare
-const nodeVersionsMap = new Map<Key, Meta>()
-const nodeVersionsList: Array<Key> = []
-let nodeVersionLatestCurrent: Key,
-	nodeVersionLatestActive: Key,
-	nodeVersionLatestMaintenance: Key
-const url =
-	'https://raw.githubusercontent.com/nodejs/Release/master/schedule.json'
+let preloaded: boolean = false,
+	nodeVersionLatestCurrent: NodeScheduleIdentifier,
+	nodeVersionLatestActive: NodeScheduleIdentifier,
+	nodeVersionLatestMaintenance: NodeScheduleIdentifier
+
+/**
+ * Fetch the metadata for the various Node.js releases from the official source.
+ * https://github.com/nodejs/Release/pull/624
+ */
+export async function preloadNodeVersions(): Promise<void> {
+	// preload
+	await preloadNodeSchedule()
+	await preloadNodeReleases()
+	preloaded = true
+
+	// helpers
+	const nodeVersionsList = getNodeScheduleIdentifiers()
+	nodeVersionLatestCurrent = last(nodeVersionsList.filter(isNodeVersionCurrent))
+	nodeVersionLatestActive = last(nodeVersionsList.filter(isNodeVersionActive))
+	nodeVersionLatestMaintenance = last(
+		nodeVersionsList.filter(isNodeVersionMaintenance)
+	)
+}
+
+// ====================================
+// FILTERING
 
 /**
  * The datetime that is used for comparisons.
@@ -53,97 +88,25 @@ let now = new Date().getTime()
 
 /**
  * Get or set the datetime that is compared against the times specified in the version metadata.
- * The only need to use this would be for testing or server purposes.
+ * Used for testing and server purposes.
  */
 export function datetime(when?: Date): number {
 	if (when != null) now = when.getTime()
 	return now
 }
 
-/**
- * Fetch the metadata for the various Node.js releases from the official source.
- * https://github.com/nodejs/Release/pull/624
- */
-export async function fetchNodeVersions(): Promise<Array<Key>> {
-	if (nodeVersionsMap.size) return nodeVersionsList
-	try {
-		// fetch node versions that have been released
-		const response = await fetch(url, {})
-		const json: Response = await response.json()
-		for (const [key, meta] of Object.entries(json)) {
-			const version = key.replace('v', '')
-			const start = new Date(meta.start)
-			const end = new Date(meta.end)
-			let maintenance: Date | undefined, lts: Date | undefined
-			if (meta.maintenance) maintenance = new Date(meta.maintenance)
-			if (meta.lts) lts = new Date(meta.lts)
-			nodeVersionsMap.set(version, {
-				version,
-				start,
-				end,
-				maintenance,
-				lts,
-				codename: meta.codename,
-			})
-		}
-		nodeVersionsList.push(...Array.from(nodeVersionsMap.keys()))
-		nodeVersionLatestCurrent = last(
-			nodeVersionsList.filter(isNodeVersionCurrent)
-		)
-		nodeVersionLatestActive = last(nodeVersionsList.filter(isNodeVersionActive))
-		nodeVersionLatestMaintenance = last(
-			nodeVersionsList.filter(isNodeVersionMaintenance)
-		)
-		return nodeVersionsList
-	} catch (err) {
-		throw new Errlop(
-			`Failed to fetch Node.js release schedule from ${url}`,
-			err
-		)
-	}
-}
-
-/**
- * Get the release metadata for the version number.
- * Requires {@link fetchNodeVersions} to have been previously awaited.
- */
-export function getNodeVersion(version: Input): Meta {
-	const meta = nodeVersionsMap.get(String(version))
-	if (!meta) {
-		if (nodeVersionsMap.size)
-			throw new Error(
-				`Unable to get the Node.js version [${JSON.stringify(
-					version
-				)}] probably because the Node.js versions have not been fetched yet.`
-			)
-		else
-			throw new Error(
-				`Unable to find the Node.js version in the cache: ${JSON.stringify(
-					version
-				)}`
-			)
-	}
-	return meta
-}
-
-/** Fetch the release metadata for the version number */
-export async function fetchNodeVersion(version: Input): Promise<Meta> {
-	await fetchNodeVersions()
-	return getNodeVersion(version)
-}
-
 /** Is the version of these versions? */
 export function isNodeVersionThese(
-	version: Input,
-	these: Array<Input>
+	version: NodeVersionInput,
+	these: Array<NodeVersionInput>
 ): boolean {
 	return Boolean(these.find((seek) => versionCompare(version, seek) === 0))
 }
 
 /** Is the version within (inclusive) of these two versions? */
 export function isNodeVersionBetween(
-	version: Input,
-	tuple: [gte: Input, lte: Input]
+	version: NodeVersionInput,
+	tuple: [gte: NodeVersionInput, lte: NodeVersionInput]
 ): boolean {
 	return (
 		versionCompare(version, tuple[0]) >= 0 &&
@@ -153,41 +116,64 @@ export function isNodeVersionBetween(
 
 /**
  * Is the version the latest active LTS release?
- * Current > Active > Maintenance > EOL
+ * Current > Active > Maintenance > EOL.
+ * Uses the Node.js Schedule API to determine this.
  */
-export function isNodeVersionLatestActive(version: Input): boolean {
+export function isNodeVersionLatestActive(version: NodeVersionInput): boolean {
+	if (nodeVersionLatestActive == null)
+		throw new Error(
+			'You must await [preloadNodeVersions] prior to using the [isNodeVersionLatestActive] filter.'
+		)
 	return versionCompare(version, nodeVersionLatestActive) === 0
 }
 
 /**
  * Is the version the latest current release?
- * Current > Active > Maintenance > EOL
+ * Current > Active > Maintenance > EOL.
+ * Uses the Node.js Schedule API to determine this.
  */
-export function isNodeVersionLatestCurrent(version: Input): boolean {
+export function isNodeVersionLatestCurrent(version: NodeVersionInput): boolean {
+	if (nodeVersionLatestCurrent == null)
+		throw new Error(
+			'You must await [preloadNodeVersions] prior to using the [isNodeVersionLatestCurrent] filter.'
+		)
 	return versionCompare(version, nodeVersionLatestCurrent) === 0
 }
 
 /**
  * Is the version the latest maintenance release?
- * Current > Active > Maintenance > EOL
+ * Current > Active > Maintenance > EOL.
+ * Uses the Node.js Schedule API to determine this.
  */
-export function isNodeVersionLatestMaintenance(version: Input): boolean {
+export function isNodeVersionLatestMaintenance(
+	version: NodeVersionInput
+): boolean {
+	if (nodeVersionLatestMaintenance == null)
+		throw new Error(
+			'You must await [preloadNodeVersions] prior to using the [isNodeVersionLatestMaintenance] filter.'
+		)
 	return versionCompare(version, nodeVersionLatestMaintenance) === 0
 }
 
 /** Is the version lesser than or equal to the seek version? */
-export function isNodeVersionLTE(version: Input, seek: Input): boolean {
+export function isNodeVersionLTE(
+	version: NodeVersionInput,
+	seek: NodeVersionInput
+): boolean {
 	return versionCompare(version, seek) <= 0
 }
 
 /** Is the version greater than or equal to the seek version? */
-export function isNodeVersionGTE(version: Input, seek: Input): boolean {
+export function isNodeVersionGTE(
+	version: NodeVersionInput,
+	seek: NodeVersionInput
+): boolean {
 	return versionCompare(version, seek) >= 0
 }
 
 /** Is the version within this range? */
 export function isNodeVersionWithinRange(
-	version: Input,
+	version: NodeVersionInput,
 	range: string
 ): boolean {
 	return withinRange(version, range)
@@ -196,9 +182,19 @@ export function isNodeVersionWithinRange(
 /**
  * Is the version an active LTS release?
  * Active LTS - New features, bug fixes, and updates that have been audited by the LTS team and have been determined to be appropriate and stable for the release line.
+ * Uses the Node.js Schedule and Release APIs to determine this.
  */
-export function isNodeVersionActive(version: Input): boolean {
-	const meta = getNodeVersion(version)
+export function isNodeVersionActive(version: NodeVersionInput): boolean {
+	if (preloaded === false)
+		throw new Error(
+			'You must await [preloadNodeVersions] prior to using the [isNodeVersionActive] filter.'
+		)
+	if (isAbsoluteVersion(version)) {
+		const lts = getNodeReleaseInformation(version).lts
+		if (typeof lts !== 'string') return false
+	}
+	const meta = getNodeScheduleInformationSafe(version)
+	if (meta === false) return false
 	const start = meta.lts
 	const end = meta.maintenance || meta.end
 	if (!start || !end) return false
@@ -209,9 +205,17 @@ export function isNodeVersionActive(version: Input): boolean {
 /**
  * Is the version a active or current release?
  * That is to say it isn't in maintenance mode, or EOL.
+ * Uses the Node.js Schedule API to determine this.
  */
-export function isNodeVersionActiveOrCurrent(version: Input): boolean {
-	const meta = getNodeVersion(version)
+export function isNodeVersionActiveOrCurrent(
+	version: NodeScheduleInput
+): boolean {
+	if (preloaded === false)
+		throw new Error(
+			'You must await [preloadNodeVersions] prior to using the [isNodeVersionActiveOrCurrent] filter.'
+		)
+	const meta = getNodeScheduleInformationSafe(version)
+	if (meta === false) return false
 	const start = meta.start
 	const end = meta.maintenance || meta.end
 	if (!start || !end) return false
@@ -219,21 +223,40 @@ export function isNodeVersionActiveOrCurrent(version: Input): boolean {
 	return activeOrCurrent
 }
 
-/** Has the version existed at some point? */
-export function isNodeVersionReleased(version: Input): boolean {
-	const meta = getNodeVersion(version)
-	const start = meta.start
-	if (!start) return false
-	const released = now >= start.getTime()
+/**
+ * Has the version existed at some point?
+ * Uses the Node.js Schedule and Release APIs to determine this.
+ */
+export function isNodeVersionReleased(version: NodeVersionInput): boolean {
+	if (preloaded === false)
+		throw new Error(
+			'You must await [preloadNodeVersions] prior to using the [isNodeVersionReleased] filter.'
+		)
+	let date: Date
+	if (isAbsoluteVersion(version)) {
+		date = getNodeReleaseInformation(version).date
+	} else {
+		const meta = getNodeScheduleInformationSafe(version)
+		if (meta === false) return false
+		date = meta.start
+	}
+	if (!date) return false
+	const released = now >= date.getTime()
 	return released
 }
 
 /**
  * Is the version a current release?
  * Current - Should incorporate most of the non-major (non-breaking) changes that land on nodejs/node master branch.
+ * Uses the Node.js Schedule and Release APIs to determine this.
  */
-export function isNodeVersionCurrent(version: Input): boolean {
-	const meta = getNodeVersion(version)
+export function isNodeVersionCurrent(version: NodeScheduleInput): boolean {
+	if (preloaded === false)
+		throw new Error(
+			'You must await [preloadNodeVersions] prior to using the [isNodeVersionCurrent] filter.'
+		)
+	const meta = getNodeScheduleInformationSafe(version)
+	if (meta === false) return false
 	const start = meta.start
 	const end = meta.lts || meta.maintenance || meta.end
 	if (!start || !end) return false
@@ -242,7 +265,7 @@ export function isNodeVersionCurrent(version: Input): boolean {
 }
 
 /** Does the version natively support ESM? */
-export function isNodeVersionESM(version: Input): boolean {
+export function isNodeVersionESM(version: NodeVersionInput): boolean {
 	return versionCompare(version, '12') >= 0
 }
 
@@ -250,9 +273,19 @@ export function isNodeVersionESM(version: Input): boolean {
  * Is the version a LTS release? Tolerates unreleased versions.
  * To exclude unreleased versions, set `released` filter to `true`.
  * Includes the historical versions of 0.8, 0.10, and 0.12 (identified by no `lts` flag and no `maintenance` flag).
+ * Uses the Node.js Schedule and Release APIs to determine this.
  */
-export function isNodeVersionLTS(version: Input): boolean {
-	const meta = getNodeVersion(version)
+export function isNodeVersionLTS(version: NodeVersionInput): boolean {
+	if (preloaded === false)
+		throw new Error(
+			'You must await [preloadNodeVersions] prior to using the [isNodeVersionLTS] filter.'
+		)
+	if (isAbsoluteVersion(version) && versionCompare(version, '1') >= 0) {
+		const lts = getNodeReleaseInformation(version).lts
+		if (typeof lts !== 'string') return false
+	}
+	const meta = getNodeScheduleInformationSafe(version)
+	if (meta === false) return false
 	const lts = Boolean(meta.lts || !meta.maintenance)
 	return lts
 }
@@ -260,9 +293,15 @@ export function isNodeVersionLTS(version: Input): boolean {
 /**
  * Is the version currently maintained?
  * Current || Active || Maintenance
+ * Uses the Node.js Schedule API to determine this.
  */
-export function isNodeVersionMaintained(version: Input): boolean {
-	const meta = getNodeVersion(version)
+export function isNodeVersionMaintained(version: NodeScheduleInput): boolean {
+	if (preloaded === false)
+		throw new Error(
+			'You must await [preloadNodeVersions] prior to using the [isNodeVersionMaintained] filter.'
+		)
+	const meta = getNodeScheduleInformationSafe(version)
+	if (meta === false) return false
 	const start = meta.start
 	const end = meta.end
 	if (!start || !end) return false
@@ -273,9 +312,17 @@ export function isNodeVersionMaintained(version: Input): boolean {
 /**
  * Is the version a maintained release or a historical LTS release? Excludes unreleased versions.
  * Includes the historical versions of 0.8, 0.10, and 0.12 (identified by no `lts` flag and no `maintenance` flag).
+ * Uses the Node.js Schedule API to determine this.
  */
-export function isNodeVersionMaintainedOrLTS(version: Input): boolean {
-	const meta = getNodeVersion(version)
+export function isNodeVersionMaintainedOrLTS(
+	version: NodeScheduleInput
+): boolean {
+	if (preloaded === false)
+		throw new Error(
+			'You must await [preloadNodeVersions] prior to using the [isNodeVersionMaintainedOrLTS] filter.'
+		)
+	const meta = getNodeScheduleInformationSafe(version)
+	if (meta === false) return false
 	const start = meta.start
 	const end = meta.end
 	if (!start || !end) return false
@@ -290,9 +337,15 @@ export function isNodeVersionMaintainedOrLTS(version: Input): boolean {
 /**
  * Is the version a maintenance release?
  * Maintenance - Critical bug fixes and security updates. New features may be added at the discretion of the LTS team - typically only in cases where the new feature supports migration to later release lines.
+ * Uses the Node.js Schedule API to determine this.
  */
-export function isNodeVersionMaintenance(version: Input): boolean {
-	const meta = getNodeVersion(version)
+export function isNodeVersionMaintenance(version: NodeScheduleInput): boolean {
+	if (preloaded === false)
+		throw new Error(
+			'You must await [preloadNodeVersions] prior to using the [isNodeVersionMaintenance] filter.'
+		)
+	const meta = getNodeScheduleInformationSafe(version)
+	if (meta === false) return false
 	const start = meta.maintenance
 	const end = meta.end
 	if (!start || !end) return false
@@ -305,25 +358,27 @@ export function isNodeVersionMaintenance(version: Input): boolean {
  * https://vercel.com/docs/serverless-functions/supported-languages?query=node%20version#defined-node.js-version
  * https://vercel.com/docs/runtimes#official-runtimes/node-js/node-js-version
  */
-export function isNodeVersionVercel(version: Input): boolean {
+export function isNodeVersionVercel(version: NodeVersionInput): boolean {
 	return (
-		versionCompare(version, '10') === 0 || versionCompare(version, '12') === 0
+		versionCompare(version, '10') === 0 ||
+		versionCompare(version, '12') === 0 ||
+		versionCompare(version, '14') === 0
 	)
 }
 
 /** The available filters to apply */
 export interface Filters {
 	/** @see isNodeVersionThese */
-	these?: Array<Input>
+	these?: Array<NodeVersionInput>
 
 	/** @see isNodeVersionBetween */
-	between?: [gte: Input, lte: Input]
+	between?: [gte: NodeVersionInput, lte: NodeVersionInput]
 
 	/** @see isNodeVersionLTE */
-	lte?: Input
+	lte?: NodeVersionInput
 
 	/** @see isNodeVersionGTE */
-	gte?: Input
+	gte?: NodeVersionInput
 
 	/** @see isNodeVersionWithinRange */
 	range?: string
@@ -369,7 +424,7 @@ export interface Filters {
 }
 
 /** Get the status of the node version? */
-export function getNodeVersionStatus(version: Input) {
+export function getNodeVersionStatus(version: NodeVersionInput) {
 	return {
 		active: isNodeVersionActive(version),
 		activeOrCurrent: isNodeVersionActiveOrCurrent(version),
@@ -388,7 +443,10 @@ export function getNodeVersionStatus(version: Input) {
 }
 
 /** Is the node version compatible with these filters? */
-export function isNodeVersion(version: Input, filters: Filters): boolean {
+export function isNodeVersion(
+	version: NodeVersionInput,
+	filters: Filters
+): boolean {
 	// with params
 	if (filters.these && !isNodeVersionThese(version, filters.these)) return false
 	if (filters.between && !isNodeVersionBetween(version, filters.between))
@@ -419,58 +477,32 @@ export function isNodeVersion(version: Input, filters: Filters): boolean {
 	return true
 }
 
-/** Filter the versions */
+/** Filter the supplied versions */
 export function filterNodeVersions(
-	versions: Array<Key>,
+	versions: Array<NodeVersionIdentifier>,
 	filters: Filters
-): Array<Key> {
+): Array<NodeVersionIdentifier> {
 	return versions.filter((version) => isNodeVersion(version, filters))
 }
 
-/** Fetch and then filter the versions */
-export async function fetchAndFilterNodeVersions(filters: Filters) {
-	await fetchNodeVersions()
-	return filterNodeVersions(nodeVersionsList, filters)
+/** Filter the significant Node.js versions */
+export function filterSignificantNodeVersions(
+	filters: Filters
+): Array<NodeVersionIdentifier> {
+	if (preloaded === false)
+		throw new Error(
+			'You must await [preloadNodeVersions] prior to using [filterSignificantNodeVersions].'
+		)
+	return filterNodeVersions(getNodeScheduleIdentifiers(), filters)
 }
 
-/** Fetch the ratified ECMAScript version at the time of a Node.js version release. */
-export async function fetchESVersionForNodeVersion(
-	nodeVersion: Input
-): Promise<string> {
-	const meta = await fetchNodeVersion(nodeVersion)
-	return getESVersion(meta.start)
-}
-
-/** Get the ratified ECMAScript version at the time of a Node.js version release. */
-export function getESVersionForNodeVersion(nodeVersion: Input): string {
-	const meta = getNodeVersion(nodeVersion)
-	return getESVersion(meta.start)
-}
-
-/**
- * For each provided Node.js version (sorted by oldest first to newest last), fetch their ECMAScript version, and return the list without duplicates.
- * @returns ECMAScript versions sorted from oldest to newest.
- */
-export async function fetchESVersionsForNodeVersions(
-	nodeVersions: Array<Input>
-): Promise<Array<string>> {
-	const versions = new Set<string>()
-	for (const nodeVersion of nodeVersions.sort(versionCompare)) {
-		versions.add(await fetchESVersionForNodeVersion(nodeVersion))
-	}
-	return Array.from(versions.values())
-}
-
-/**
- * For each provided Node.js version (sorted by oldest first to newest last), fetch their ECMAScript version, and return the list without duplicates.
- * @returns ECMAScript versions sorted from oldest to newest.
- */
-export function getESVersionsForNodeVersions(
-	nodeVersions: Array<Input>
-): Array<string> {
-	const versions = new Set<string>()
-	for (const nodeVersion of nodeVersions.sort(versionCompare)) {
-		versions.add(getESVersionForNodeVersion(nodeVersion))
-	}
-	return Array.from(versions.values())
+/** Filter the absolute Node.js versions */
+export function filterAbsoluteNodeVersions(
+	filters: Filters
+): Array<NodeVersionIdentifier> {
+	if (preloaded === false)
+		throw new Error(
+			'You must await [preloadNodeVersions] prior to using [filterAbsoluteNodeVersions].'
+		)
+	return filterNodeVersions(getNodeReleaseIdentifiers(), filters)
 }