+++ HISTORY.md
@@ -1,5 +1,12 @@
 # History
 
+## v2.0.0 2021 July 27
+
+-   Rewrote as `@bevry/nodejs-versions`
+-   Now supports both significant and absolute versions
+-   API is now simpler, with the various `fetch*` calls now a single `preloadNodeVersions()` call
+-   Updated dependencies, [base files](https://github.com/bevry/base), and [editions](https://editions.bevry.me) using [boundation](https://github.com/bevry/boundation)
+
 ## v1.1.0 2020 November 12
 
 -   Add `fetchESVersionForNodeVersion`, `getESVersionForNodeVersion`, `fetchESVersionForNodeVersions`, `getESVersionForNodeVersions`